% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayOfYear.R
\name{DayOfYear}
\alias{DayOfYear}
\title{Day of year}
\usage{
DayOfYear(DateTime)
}
\arguments{
\item{DateTime}{DateTime object}
}
\description{
This function returns a continuous the day of year value (as integer value 1:365) for a given date-time in "POSIXlt" "POSIXct" format.
}
\examples{

#Calculating day of year for now

DayOfYear(Sys.time())

}
\keyword{DOY,}
\keyword{Day}
\keyword{of}
\keyword{year}
