% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Declination.R
\name{Declination}
\alias{Declination}
\title{Declination Angle}
\usage{
Declination(DOY)
}
\arguments{
\item{DOY}{Day of year}
}
\description{
This function calculates solar declination angle for a given day of year.
}
\examples{

#Calculating solar declination angle for 365 day of the year

DOY <- 1:365

delta <- Declination(DOY)

plot(DOY, delta)

}
\keyword{Angle}
\keyword{Declination}
\keyword{Solar}
