% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraterrestrial.R
\name{Extraterrestrial}
\alias{Extraterrestrial}
\title{Solar Extraterrestrial Radiation}
\usage{
Extraterrestrial(DOY)
}
\arguments{
\item{DOY}{Day of year}
}
\description{
This function calculates solar extraterrestrial radiation (in W/m2) for a given day of year.
}
\examples{

#Calculating solar extraterrestrial radiation for 365 day of the year

DOY <- 1:365

Sextr <- Extraterrestrial(DOY)

plot(DOY, Sextr)

}
\keyword{Extraterrestrial}
\keyword{Radiation}
\keyword{Solar}
