% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCL.R
\name{getCL}
\alias{getCL}
\title{Center line column of an image}
\usage{
getCL(file)
}
\arguments{
\item{file}{a character string, path to the JPEF file}
}
\value{
a three-column matrix of red, green blue bands of the center line
}
\description{
This function returns the R,G,B vectors extracted from the center-line of a JPEG file
}
\examples{

f <- system.file(package = 'xROI', 'dukehw.jpg')
cli <- getCL(f)

}
\keyword{CLI}
\keyword{Center-line}
\keyword{image}
