% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCLArray.R
\name{getCLArray}
\alias{getCLArray}
\title{Center line arrary of an array of image}
\usage{
getCLArray(files)
}
\arguments{
\item{files}{a vector of character strings, paths to the JPEF files}
}
\value{
A 3D array. The center line image as an array (NxHx3), where N is number of files, and H is the height of an image in pixels.
}
\description{
This function returns CLI array for vector of JPEG files
}
\examples{

f <- system.file(package = 'xROI', 'dukehw.jpg')
cli <- getCL(f)

}
\keyword{CLI}
\keyword{Center-line}
\keyword{image}
