% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServer.R
\name{getServer}
\alias{getServer}
\title{Server function}
\usage{
getServer(exdir, inputDir = NULL)
}
\arguments{
\item{exdir}{path to the example directory}

\item{inputDir}{path to the input directory}
}
\description{
This funciton constructs the server functions of the shiny app.
}
\examples{
exampleDir <- system.file('example', package = "xROI")
server <- getServer(exampleDir)

}
\keyword{App}
\keyword{Server}
\keyword{Shiny}
