% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettmpdir.R
\name{gettmpdir}
\alias{gettmpdir}
\title{Path to the TEMP directory}
\usage{
gettmpdir()
}
\value{
the path to the system temporary directory
}
\description{
This function returns the path to the TEMP directory
}
\examples{
p <- gettmpdir()

}
\keyword{directory}
\keyword{temporary}
