% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarConstant.R
\docType{data}
\name{SolarConstant}
\alias{SolarConstant}
\title{Solar Constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
SolarConstant
}
\description{
This constant value returns solar constant in Watt per meter squared
}
\examples{

#Printing Solar Constant

print(SolarConstant)
}
\keyword{Constant}
\keyword{Solar}
