% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sunrise.R
\name{Sunrise}
\alias{Sunrise}
\title{Sunrise Time}
\usage{
Sunrise(DOY, Lat)
}
\arguments{
\item{DOY}{Day of year}

\item{Lat}{Latitude (in degrees)}
}
\description{
This function estimates sunrise time (in continuous hour values) for a given day of year and latitude.
}
\examples{

#Calculating sunrise time for 365 day of the year for 45 degree latitude

DOY <- 1:365

Lat = 45

sunrise <- Sunset(DOY, Lat)

plot(DOY, sunrise)

}
\keyword{Sunrise}
