% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeROI.R
\name{writeROI}
\alias{writeROI}
\title{Write ROI list file}
\usage{
writeROI(ROIList, roifilepath)
}
\arguments{
\item{ROIList}{ROI List variable to be written}

\item{roifilepath}{path to the ROI file}
}
\description{
This function writes the ROI list file on a disk space.
}
\examples{

f <- system.file(package = 'xROI', 'example/ROI/example_DB_0001_roi.csv')
roi <- parseROI(f)
writeROI(roi, 'roi.csv')

}
\keyword{ROI}
\keyword{ROIList}
\keyword{Write}
